+++ obackup.sh
@@ -3,7 +3,7 @@
 ###### Remote (or local) backup script for files & databases
 ###### (L) 2013 by Orsiris "Ozy" de Jong (www.netpower.fr)
 OBACKUP_VERSION=1.84preRC3-MSYS-compatible
-OBACKUP_BUILD=1010201301
+OBACKUP_BUILD=1110201301
 
 DEBUG=no
 SCRIPT_PID=$$
@@ -93,7 +93,7 @@ function TrapQuit
 	if type -p pkill > /dev/null 2>&1
 	then
 		pkill -TERM -P $$
-	elif [ "$OSTYPE" == "msys" ]
+	elif [ "$LOCAL_OS" == "msys" ]
 	then
 		## This is not really a clean way to get child process pids, especially the tail -n +2 which resolves a strange char apparition in msys bash
 		for pid in $(ps -a | awk '{$1=$1}$1' | awk '{print $1" "$2}' | grep " $$$" | awk '{print $1}' | tail -n +2)
@@ -174,6 +174,7 @@ function CleanUp
 {
 	if [ "$DEBUG" != "yes" ]
 	then
+		rm -f "$RUN_DIR/obackup_remote_os_$SCRIPT_PID"
 		rm -f "$RUN_DIR/obackup_dblist_$SCRIPT_PID"
 		rm -f "$RUN_DIR/obackup_local_sql_storage_$SCRIPT_PID"
         	rm -f "$RUN_DIR/obackup_local_file_storage_$SCRIPT_PID"
@@ -283,12 +284,55 @@ function CheckEnvironment
 	fi
 }
 
+function GetOperatingSystem
+{
+        LOCAL_OS_VAR=$(uname -spio)
+        if [ "$REMOTE_SYNC" == "yes" ]
+        then
+                eval "$SSH_CMD uname -spio > $RUN_DIR/obackup_remote_os_$SCRIPT_PID 2>&1 &"
+                REMOTE_OS_VAR=$(cat $RUN_DIR/obackup_remote_os_$SCRIPT_PID)
+        fi
+
+        case $LOCAL_OS_VAR in
+                "Linux"*)
+                LOCAL_OS="Linux"
+                ;;
+                "FreeBSD"*)
+                LOCAL_OS="FreeBSD"
+                ;;
+                "MINGW32"*)
+                LOCAL_OS="msys"
+                ;;
+                *)
+                LogError "Running on >> $LOCAL_OS_VAR << not supported. Please report to the author."
+                exit 1
+                ;;
+        esac
+
+        case $REMOTE_OS_VAR in
+                "Linux"*)
+                REMOTE_OS="Linux"
+                ;;
+                "FreeBSD"*)
+                REMOTE_OS="FreeBSD"
+                ;;
+                "MINGW32"*)
+                REMOTE_OS="msys"
+                ;;
+                "")
+                ;;
+                *)
+                LogError "Running on remote >> $REMOTE_OS_VAR << not supported. Please report to the author."
+                exit 1
+        esac
+}
+
 # Waits for pid $1 to complete. Will log an alert if $2 seconds exec time exceeded unless $2 equals 0. Will stop task and log alert if $3 seconds exec time exceeded.
 function WaitForTaskCompletion
 {
         soft_alert=0
         SECONDS_BEGIN=$SECONDS
-                if [ "$OSTYPE" == "msys" ]
+                if [ "$LOCAL_OS" == "msys" ]
         then
         	PROCESS_TEST="ps -a | awk '{\$1=\$1}\$1' | awk '{print \$1}' | grep $1"
         else
@@ -534,11 +578,11 @@ function CheckConnectivityRemoteHost
 {
 	if [ "$REMOTE_HOST_PING" != "no" ] && [ "$REMOTE_BACKUP" != "no" ]
 	then
-		if [ "$OSTYPE" == "msys" ]
+		if [ "$LOCAL_OS" == "msys" ]
 		then
-			ping $REMOTE_HOST -n 2 > /dev/null 2>&1
+			ping -n 2 $REMOTE_HOST > /dev/null 2>&1
 		else
-			ping $REMOTE_HOST -c 2 > /dev/null 2>&1
+			ping -c 2 $REMOTE_HOST > /dev/null 2>&1
 		fi
 		if [ $? != 0 ]
 		then
@@ -557,11 +601,11 @@ function CheckConnectivity3rdPartyHosts
                 IFS=$' \t\n'
 		for i in $REMOTE_3RD_PARTY_HOSTS
 		do
-			if [ "$OSTYPE" == "msys" ]
+			if [ "$LOCAL_OS" == "msys" ]
 			then
-				ping $i -n 2 > /dev/null 2>&1
+				ping -n 2 $i > /dev/null 2>&1
 			else
-				ping $i -c 2 > /dev/null 2>&1
+				ping -c 2 $i > /dev/null 2>&1
 			fi
 			if [ $? != 0 ]
 			then
@@ -1003,7 +1047,7 @@ function Init
 	MAIL_ALERT_MSG="Warning: Execution of obackup instance $BACKUP_ID (pid $SCRIPT_PID) as $LOCAL_USER@$LOCAL_HOST produced errors."
 
 	## If running Msys, find command of windows is used instead of msys one
-	if [ "$OSTYPE" == "msys" ]
+	if [ "$LOCAL_OS" == "msys" ]
 	then
 		FIND_CMD=$(dirname $BASH)/find
 	else